backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic==2.9.0
openai==1.3.7
httpx==0.24.1
pdfplumber==0.11.7
